Move Progress back to original location.
