+++ lib/console/event.rb
@@ -5,4 +5,3 @@
 
 require_relative 'event/spawn'
 require_relative 'event/failure'
-require_relative 'event/progress'
