+++ test/console/event/progress.rb
@@ -3,12 +3,12 @@
 # Released under the MIT License.
 # Copyright, 2020-2022, by Samuel Williams.
 
-require 'console/event/progress'
+require 'console/progress'
 
 require 'console'
 require 'console/capture'
 
-describe Console::Event::Progress do
+describe Console::Progress do
 	let(:capture) {Console::Capture.new}
 	let(:logger) {Console::Logger.new(capture)}
 	let(:progress) {logger.progress("My Measurement", 100)}
