+++ test/console/terminal/formatter/progress.rb
@@ -1,5 +1,5 @@
 require 'console/terminal/formatter/progress'
-require 'console/event/progress'
+require 'console/progress'
 require 'console/output/null'
 require 'console/terminal'
 
@@ -9,12 +9,12 @@ describe Console::Terminal::Formatter::Progress do
 	let(:formatter) {subject.new(terminal)}
 	let(:output) {Console::Output::Null.new}
 	
-	let(:event) do
-		Console::Event::Progress.new(output, self, 10)
+	let(:progress) do
+		Console::Progress.new(output, self, 10)
 	end
 	
 	it "can format failure events" do
-		formatter.format(event.to_hash, buffer)
+		formatter.format(progress.to_hash, buffer)
 		
 		expect(buffer.string).to be =~ /0.00%/
 	end
